% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\title{Interpolate the locations of values within an ordered vector viewed from left to right}
\usage{
locate(x, vec, ...)
}
\arguments{
\item{x}{vector of values (in any order) including sortable character types}

\item{vec}{ordered vector (ascending or descending) of same type as \code{x}}

\item{...}{other arguments for \code{sapply}}
}
\value{
vector of indices ranging from 0 to the length of \code{vec}
\itemize{
\item{0 indices indicate \code{x} values less than the left most value of \code{vec}}
\item{For ascending vectors, the index, i, of x is found where \code{vex[i] <= x < vec[i+1]} with an open right-hand end}
\item{For descending vectors, the index, i, of x is found where \code{vex[i] >= x > vec[i+1]} with an open right-hand end}
}
}
\description{
Inpsired by \href{http://www.numerical.recipes/}{locate() from numerical recipes},
  \code{\link[base]{findInterval}} and
  \href{https://www.l3harrisgeospatial.com/docs/value_locate.html}{VALUE_LOCATE() from IDL}.
}
\examples{
\dontrun{
  x <- c(13, 7, 22)
  locate(x, 10:19)
  locate(x, 19:10)
  locate(c("zoo", "dog", "fish", "cat"), letters)
  locate(c("zoo", "dog", "fish", "cat"), letters, USE.NAMES = FALSE)
}
}
